+++ src/frontend/src/binder/select.rs
@@ -715,7 +715,7 @@ fn data_type_to_alias(data_type: &AstDataType) -> Option<String> {
         AstDataType::Array(ty) => return data_type_to_alias(ty),
         AstDataType::Custom(ty) => format!("{}", ty),
         AstDataType::Struct(_) | AstDataType::Map(_) => {
-            // Note: Postgres doesn't have maps and anonymous structs
+            // It doesn't bother to derive aliases for these types.
             return None;
         }
     };
